Tpt inheritance, LLBLGen model synced Weather service implemented for all ORMs

+++ src/NHibernateSample/Services/AnimalService.cs
@@ -51,7 +51,7 @@ namespace MTech.NHibernateSample
         public IList<Animal> GetByType(AnimalType type)
         {
             return _session.Query<Animal>()
-                .Where(animal => animal.Type == type)
+                .Where(animal => (int)animal.Type == (int)type)
                 .ToList();
         }
